refactor(automation): split prompts.py (1368 LOC) into phase-grouped package
Converts hephaestus/automation/prompts.py into a hephaestus/automation/prompts/
package with 8 phase-grouped submodules (implementation, planning, advise,
follow_up, pr_review, address_review, plus _shared helpers and
_strict_rubric constants). Top-level __init__.py re-exports every public
symbol so all 8 existing importers stay unchanged.

Pure move-and-re-export. No behavior change. Every string template is
character-for-character preserved.

Part of #607.
