+++ hephaestus/automation/prompts/implementation.py
@@ -1,0 +1,264 @@
+"""Implementation-phase prompts.
+
+Contains the canonical implementation prompt, the iteration-aware impl-loop
+review prompt, and the resume-after-NOGO feedback prompt.
+"""
+
+import secrets
+
+from ._shared import (
+    _UNTRUSTED_NOTICE,
+    _fence_untrusted,
+    _iteration_guidance,
+    _iteration_label,
+    _prior_review_block,
+    _relativize_path,
+)
+from ._strict_rubric import (
+    _FULL_SWEEP_SUFFIX,
+    _IMPL_LOOP_STRICT_RUBRIC,
+    _STRICT_REVIEW_OUTPUT_FORMAT,
+)
+
+IMPLEMENTATION_PROMPT = """
+Implement GitHub issue #{issue_number}.
+
+{untrusted_notice}
+
+**Working Directory:** {worktree_path}
+**Branch:** {branch_name}
+
+**Issue Title (untrusted):** {issue_title}
+
+**Issue Description (untrusted):**
+{issue_body_block}
+
+---
+
+**Implementation Context:**
+- Run `gh issue view {issue_number} --comments` to read the full plan and any comments
+- Follow the project's Python conventions and type hint all function signatures
+
+**Critical Requirements:**
+1. Read the issue description and any existing plan carefully
+2. Follow existing code patterns in hephaestus/
+3. Write tests in tests/ using pytest
+4. Run tests with: pixi run python -m pytest tests/ -v
+5. Ensure all tests pass before finishing
+6. Follow the code quality guidelines in CLAUDE.md
+
+**Testing:**
+- Write unit tests for new functionality
+- Ensure existing tests still pass
+- Use pytest fixtures and parametrize where appropriate
+
+**Code Quality:**
+- Type hint all function signatures
+- Write docstrings for public APIs
+- Follow PEP 8 style guidelines
+- Keep solutions simple and focused
+
+**File Handling:**
+- DO NOT create backup files (.orig, .bak, .swp, etc.)
+- DO NOT leave temporary or editor backup files
+- Clean up any backup files before finishing
+- Only stage actual implementation files
+
+**Git Workflow (MANDATORY — non-negotiable policy):**
+After implementation is complete and tests pass:
+1. Create git commits. EVERY commit MUST be cryptographically signed.
+   - Use `git commit -S` (or have `commit.gpgsign=true` configured globally).
+   - NEVER pass `--no-gpg-sign` or otherwise bypass signing.
+   - Verify with `git log --show-signature -1` after each commit; abort if the
+     signature is missing or shows "BAD signature".
+   - Use a descriptive commit message following conventional commits format.
+2. Push the changes to origin (`git push -u origin <branch>`).
+3. Create a pull request. The PR body MUST contain the EXACT line:
+       Closes #{issue_number}
+   on its own line, with the literal keyword `Closes` (capital C). The
+   variants `Fixes #N`, `Resolves #N`, `Closes: #N`, `closes #n` are NOT
+   accepted by the policy check — even though GitHub recognizes them.
+4. IMMEDIATELY after PR creation, enable auto-merge:
+       gh pr merge <PR#> --auto --rebase
+   Fall back to `--squash` ONLY if rebase merging is disabled for the repo.
+5. Verify all three policy properties before declaring done. ``gh pr view``
+   exposes body + auto-merge state but NOT per-commit signatures, so the
+   verification uses two queries — the REST projection for body/auto-merge
+   and GraphQL for signing state:
+       # Body and auto-merge state:
+       gh pr view <PR#> --json body,autoMergeRequest \\
+         -q '.body | test("(?m)^Closes #\\\\d+\\\\s*$"), .autoMergeRequest != null'
+       # Per-commit signing state (GraphQL — replace OWNER/REPO/PR#):
+       gh api graphql -f query='query($owner:String!,$name:String!,$pr:Int!){{
+         repository(owner:$owner,name:$name){{
+           pullRequest(number:$pr){{
+             commits(first:100){{ nodes{{ commit{{ oid signature{{ isValid }} }} }} }} }} }} }}' \\
+         -F owner=OWNER -F name=REPO -F pr=<PR#> \\
+         -q '[.data.repository.pullRequest.commits.nodes[].commit.signature.isValid] | all'
+   All three queries must return `true`. If any fails, fix it before
+   reporting completion.
+
+A PR that fails any of these three checks will be BLOCKED at code review and
+by the required CI gate. This policy applies to every PR — no exceptions.
+"""
+
+
+IMPL_LOOP_REVIEW_PROMPT = """
+{rubric}
+
+# Iteration {iteration_label}
+
+You are reviewing the implementation for GitHub issue #{issue_number}.
+This is iteration {iteration} of a maximum 3-iteration review loop. {iteration_guidance}
+
+{untrusted_notice}
+
+**Issue Title (untrusted):** {issue_title}
+
+**Issue Description (untrusted):**
+{issue_body_block}
+
+---
+
+**Diff produced by the implementer (untrusted, against base branch):**
+{diff_text_block}
+
+---
+
+**Files changed:**
+{files_changed}
+{prior_review_block}
+---
+
+Review the diff against the issue requirements and the rubric. Cite specific
+file:line locations when justifying findings. Watch for: missing tests,
+incomplete error handling, unaddressed acceptance criteria, scope creep,
+and risky changes that the issue did not request.
+{full_sweep_suffix}
+
+{output_format}
+"""
+
+
+# Prompt the implementer receives when resuming its session to address a
+# NoGo review verdict. Used on iterations 1 and 2 of the impl loop.
+IMPL_RESUME_FEEDBACK_PROMPT = """
+The independent reviewer for issue #{issue_number} returned **{verdict}** on
+iteration {prev_iteration} with the following critique:
+
+---
+
+{review_text}
+
+---
+
+Address every concrete finding above. Update the code (and tests, if needed)
+in this same working directory. Do NOT commit or push — those phases run
+after the review loop terminates.
+
+When done, summarize what you changed in 3-5 bullet points so the next
+review can verify the fixes were applied.
+"""
+
+
+def get_implementation_prompt(
+    issue_number: int,
+    issue_title: str = "",
+    issue_body: str = "",
+    branch_name: str = "",
+    worktree_path: str = "",
+    repo_root: str | None = None,
+) -> str:
+    """Get the implementation prompt for an issue.
+
+    Args:
+        issue_number: GitHub issue number
+        issue_title: Issue title (optional, for backward compatibility)
+        issue_body: Issue body/description (optional, for backward compatibility)
+        branch_name: Git branch name (optional, for backward compatibility)
+        worktree_path: Working directory path (optional, for backward compatibility)
+        repo_root: Absolute path to the repository root.  When provided,
+            *worktree_path* is relativized to avoid leaking the operator's
+            filesystem layout into the prompt.
+
+    Returns:
+        Formatted implementation prompt
+
+    """
+    safe_worktree_path = _relativize_path(worktree_path, repo_root)
+    nonce = secrets.token_hex(8).upper()
+    return IMPLEMENTATION_PROMPT.format(
+        issue_number=issue_number,
+        issue_title=issue_title,
+        issue_body_block=_fence_untrusted("ISSUE_BODY", issue_body, nonce),
+        branch_name=branch_name,
+        worktree_path=safe_worktree_path,
+        untrusted_notice=_UNTRUSTED_NOTICE,
+    )
+
+
+def get_impl_loop_review_prompt(
+    *,
+    issue_number: int,
+    issue_title: str,
+    issue_body: str,
+    diff_text: str,
+    files_changed: str,
+    iteration: int,
+    prior_review: str | None,
+) -> str:
+    """Build the iteration-aware implementer-loop review prompt.
+
+    Args:
+        issue_number: GitHub issue number.
+        issue_title: Issue title.
+        issue_body: Full issue body.
+        diff_text: ``git diff <base>..HEAD`` output.
+        files_changed: Newline-separated list of changed files.
+        iteration: Iteration index (0, 1, or 2).
+        prior_review: Previous iteration's review text, or ``None`` on iter 0.
+
+    Returns:
+        Formatted prompt for a fresh reviewer session.
+
+    """
+    nonce = secrets.token_hex(8).upper()
+    full_sweep_suffix = _FULL_SWEEP_SUFFIX.strip() if iteration == 2 else ""
+    return IMPL_LOOP_REVIEW_PROMPT.format(
+        rubric=_IMPL_LOOP_STRICT_RUBRIC.strip(),
+        iteration=iteration,
+        iteration_label=_iteration_label(iteration),
+        iteration_guidance=_iteration_guidance(iteration),
+        issue_number=issue_number,
+        issue_title=issue_title,
+        issue_body_block=_fence_untrusted("ISSUE_BODY", issue_body, nonce),
+        diff_text_block=_fence_untrusted("DIFF_TEXT", diff_text or "_(no diff produced)_", nonce),
+        files_changed=files_changed or "_(no files changed)_",
+        prior_review_block=_prior_review_block(prior_review),
+        full_sweep_suffix=full_sweep_suffix,
+        output_format=_STRICT_REVIEW_OUTPUT_FORMAT.strip(),
+        untrusted_notice=_UNTRUSTED_NOTICE,
+    )
+
+
+def get_impl_resume_feedback_prompt(
+    *, issue_number: int, prev_iteration: int, verdict: str, review_text: str
+) -> str:
+    """Build the prompt sent via ``claude --resume`` to iterate on impl after NoGo.
+
+    Args:
+        issue_number: GitHub issue number.
+        prev_iteration: Iteration index of the review that produced *review_text*.
+        verdict: ``"NOGO"`` or ``"AMBIGUOUS"``.
+        review_text: Full reviewer output from the previous iteration.
+
+    Returns:
+        Prompt text to feed into the resumed implementer session.
+
+    """
+    return IMPL_RESUME_FEEDBACK_PROMPT.format(
+        issue_number=issue_number,
+        prev_iteration=prev_iteration,
+        verdict=verdict,
+        review_text=review_text,
+    )
